supervisor/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
confluent-kafka==2.3.0
idna==3.4
mysql-connector-python==8.2.0
protobuf==4.21.12
redis==5.0.1
requests==2.31.0
urllib3==2.1.0
pika==1.3.2